fix(*): TypeScript conditional require based on webpack.DefinePlugin env var
Some modules are used in development that should not be part of the final bundle (such as ngrx/devtools, loggers ...), they only add weight.
Thanks to webpack.DefinePlugin, those modules are now dropped after compilation step.
This isn't as straightforward as with webpack/Babel/ES6 (because of TypeScript typing system and Angular decorators).

At the end you gain 24.5KB (it could be much more if you had a lot of dependencies or polymorphic bundles)

Fixes #1